Modulo order modificado para funcionar según cliente o consumidor final

+++ backend/src/modules/order/order.service.ts
@@ -9,7 +9,6 @@ import { Product } from 'src/entities/products/product.entity';
 import { ProductsOrder } from 'src/entities/product-order.entity';
 import { OrderQuery } from './orders.query';
 import { Transaccion } from 'src/entities/transaction.entity';
-import { OrderStatus } from 'src/enum/orderStatus.enum';
 
 @Injectable()
 export class OrderService {
@@ -35,25 +34,19 @@ export class OrderService {
     }
 
     async getOrdersByUserId(id: string) {
-        return await this.orderQuery.getOrdersByUserId(id)
+        return await this.orderQuery.getOrdersByUserId(id);
     }
 
-    async createOrder(
-      userId: string, 
-      productsInfo: ProductInfo[], 
-      address: string | undefined, 
-      discount: number | undefined, 
-      deliveryDate: Date | undefined
-    ) {
+    async createOrder(userId: string, productsInfo: ProductInfo[], address: string | undefined, account: boolean) {
         let total = 0; 
         let createdOrder;
   
         const user = await this.userRepository.findOneBy({ id: userId, isDeleted: false });   
-        if (!user) throw new BadRequestException(`User not found. ID: ${userId}`);
+        if (!user) throw new BadRequestException(`Usuario no encontrado. ID: ${userId}`);
     
         await Promise.all(productsInfo.map(async (product)=> {
             const foundProduct = await this.productRepository.findOneBy({ id: product.id });
-            if (!foundProduct) throw new BadRequestException(`Product not found. ID: ${product.id}`);
+            if (!foundProduct) throw new BadRequestException(`Producto no encontrado. ID: ${product.id}`);
             if (foundProduct.stock <= 0) throw new BadRequestException(`Producto sin stock. ID: ${foundProduct.id}`);
         }));
   
@@ -71,7 +64,6 @@ export class OrderService {
                 total += ((foundProduct.price * product.quantity) * (1 - foundProduct.discount));
                 
                 const productsOrder = transactionalEntityManager.create(ProductsOrder, {
-                    
                     product: foundProduct, 
                     order: newOrder,
                     quantity: product.quantity
@@ -80,25 +72,22 @@ export class OrderService {
                 await transactionalEntityManager.save(ProductsOrder, productsOrder);
             }));
   
-            if (discount) total *= (1 - discount);
-  
             const orderDetail = transactionalEntityManager.create(OrderDetail, {
                 totalPrice: Number(total.toFixed(2)),
                 order: newOrder,
-                discount: discount || 0,
-                addressDelivery: address || 'Tienda',
-                deliveryDate
+                addressDelivery: address || 'Retiro en local',
             });
   
             await transactionalEntityManager.save(OrderDetail, orderDetail);
   
             await transactionalEntityManager.save(Transaccion, {
-                status: "Solicitado",
+                status: account ? 'En preparación' : 'Pendiente de pago',
                 timestamp: new Date(),
                 orderdetail: orderDetail
             });
         });
   
+        delete createdOrder.user.password;
         return createdOrder;
     }
 
@@ -119,9 +108,26 @@ export class OrderService {
 
         return { HttpCode: 200 };
     }
+
+    async MercadoPagoUpdate(id: string) {
+        console.log('tamo aca')
+        const foundOrder = await this.orderRepository.findOne(
+            { where: { id }, 
+            relations: { orderDetail: { transactions: true } } 
+        });
+        if (!foundOrder) throw new BadRequestException(`Orden no encontrada. ID: ${id}`);
+
+        await this.orderRepository.update(id, { status: 'Pagado' });
+        await this.transactionRepository.update({ id: foundOrder.orderDetail.transactions.id }, { status: 'En preparación' });
+
+        return { HttpCode: 200 };
+    }
   
     async deleteOrder(id: string) {
-      return await this.orderQuery.deleteOrder(id);
+        const foundOrder = await this.orderQuery.getOrderById(id);
+        if(!foundOrder) throw new NotFoundException(`Orden no encontrada. ID: ${id}`);   
+      
+        return await this.orderQuery.deleteOrder(id);
     }
     
     async updateStock(id: string) {
